+++ packages/ckeditor5-image/src/imageresize.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
@@ -57,7 +57,7 @@ export default class ImageResize extends Plugin {
 
 					modelElement: data.item,
 					viewElement: widget,
-					downcastWriter: conversionApi.writer,
+					editor,
 
 					getHandleHost( domWidgetElement ) {
 						return domWidgetElement.querySelector( 'img' );