+++ packages/ckeditor5-image/src/imagetextalternative/imagetextalternativeui.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
@@ -78,7 +78,9 @@ export default class ImageTextAlternativeUI extends Plugin {
 
 			view.bind( 'isEnabled' ).to( command, 'isEnabled' );
 
-			this.listenTo( view, 'execute', () => this._showForm() );
+			this.listenTo( view, 'execute', () => {
+				this._showForm();
+			} );
 
 			return view;
 		} );