+++ packages/ckeditor5-image/src/imagetextalternative/ui/textalternativeformview.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
@@ -110,7 +110,7 @@ export default class TextAlternativeFormView extends View {
 			attributes: {
 				class: [
 					'ck',
-					'ck-text-alternative-form',
+					'ck-text-alternative-form'
 				],
 
 				// https://github.com/ckeditor/ckeditor5-image/issues/40