+++ packages/ckeditor5-image/src/imagetoolbar.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
@@ -50,7 +50,7 @@ export default class ImageToolbar extends Plugin {
 		widgetToolbarRepository.register( 'image', {
 			ariaLabel: t( 'Image toolbar' ),
 			items: editor.config.get( 'image.toolbar' ) || [],
-			getRelatedElement: getSelectedImageWidget,
+			getRelatedElement: getSelectedImageWidget
 		} );
 	}
 }