+++ packages/ckeditor5-image/tests/imagestyle/imagestyleediting.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
@@ -53,7 +53,7 @@ describe( 'ImageStyleEditing', () => {
 						styles: [
 							{ name: 'fullStyle', title: 'foo', icon: 'object-center', isDefault: true },
 							{ name: 'sideStyle', title: 'bar', icon: 'object-right', className: 'side-class' },
-							{ name: 'dummyStyle', title: 'baz', icon: 'object-dummy', className: 'dummy-class' },
+							{ name: 'dummyStyle', title: 'baz', icon: 'object-dummy', className: 'dummy-class' }
 						]
 					}
 				} )