+++ packages/ckeditor5-image/tests/imagetoolbar.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2019, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
@@ -61,7 +61,7 @@ describe( 'ImageToolbar', () => {
 		global.document.body.appendChild( editorElement );
 
 		return ClassicEditor.create( editorElement, {
-			plugins: [ ImageToolbar ],
+			plugins: [ ImageToolbar ]
 		} )
 			.then( editor => {
 				expect( editor.plugins.get( ImageToolbar )._toolbar ).to.be.undefined;